+++ internal/cmd/root.go
@@ -16,7 +16,7 @@ var (
 	cfgFile  string
 	logLevel string
 	cfg      *config.Config
-	Version  = "0.5.1"
+	Version  = "0.5.2"
 )
 
 var rootCmd = &cobra.Command{